fix(evalforge-skill-gate): skip the run when the PR author cannot be resolved
The author lookup was the one unguarded network call in runGate, so a GitHub error
escaped to index.ts's catch and became setFailed — a red check regardless of
`blocking`. That broke the one promise soak mode makes. Verified against the built
bundle with an invalid token: exit code was 1, now 0.

Skipping is the right answer rather than failing: the gate's job here is deciding
whether to run at all, so an unresolvable author means don't run — no cost, no
verdict, no broken PR. It also no longer misreports the cause as "author is not a
@wix.com address" when the lookup simply failed.

draft-tags.ts becomes pr-lookups.ts and holds both GitHub PR lookups, which share
the property worth stating once: swallow the error, return the safe answer.
